refactor: move reconcileClustersForRegions from clusters mgr to dynamic scaleup manager
